refactor: remove nested compat bridges
