--- rustfs/src/admin/service/storage_compat.rs
@@ -1,30 +1,0 @@
-// Copyright 2024 RustFS Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub(crate) use super::super::storage_compat::{
-    Error, STORAGE_CLASS_SUB_SYS, get_global_notification_sys, read_admin_config, read_admin_config_without_migrate,
-    save_admin_config, set_global_storage_class,
-};
-
-#[cfg(test)]
-pub(crate) use super::super::storage_compat::init_admin_config_defaults;
-
-#[cfg(test)]
-pub(crate) mod metadata {
-    pub(crate) use super::super::super::storage_compat::metadata::{BUCKET_LIFECYCLE_CONFIG, BUCKET_REPLICATION_CONFIG};
-}
-
-pub(crate) mod storageclass {
-    pub(crate) use super::super::super::storage_compat::storageclass::lookup_config;
-}
