+++ rustfs/src/app/context.rs
@@ -21,7 +21,6 @@ mod global;
 mod handles;
 mod interfaces;
 mod startup;
-pub(crate) mod storage_compat;
 
 pub use compat::*;
 pub use global::*;
