+++ rustfs/src/app/context/compat.rs
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::super::storage_compat::ECStore;
+use super::super::storage_compat::EndpointServerPools;
+use super::super::storage_compat::TierConfigMgr;
+use super::super::storage_compat::metadata_sys::BucketMetadataSys;
+use super::super::storage_compat::new_object_layer_fn;
 use super::global::{AppContext, get_global_app_context};
 use super::handles::{
     default_bucket_metadata_interface, default_endpoints_interface, default_kms_runtime_interface,
     default_server_config_interface, default_tier_config_interface,
 };
-use super::storage_compat::ECStore;
-use super::storage_compat::EndpointServerPools;
-use super::storage_compat::TierConfigMgr;
-use super::storage_compat::metadata_sys::BucketMetadataSys;
-use super::storage_compat::new_object_layer_fn;
 #[cfg(test)]
 use crate::config::RustFSBufferConfig;
 use rustfs_config::server_config::Config;
@@ -119,10 +119,10 @@ fn resolve_buffer_config_with(
 
 #[cfg(test)]
 mod tests {
-    use super::super::storage_compat::Endpoint;
-    use super::super::storage_compat::init_local_disks;
-    use super::super::storage_compat::new_object_layer_fn;
-    use super::super::storage_compat::{Endpoints, PoolEndpoints};
+    use super::super::super::storage_compat::Endpoint;
+    use super::super::super::storage_compat::init_local_disks;
+    use super::super::super::storage_compat::new_object_layer_fn;
+    use super::super::super::storage_compat::{Endpoints, PoolEndpoints};
     use super::*;
     use crate::app::context::global::AppContextTestInterfaces;
     use crate::app::context::handles::{default_notify_interface, default_region_interface};
