+++ rustfs/src/app/context/interfaces.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::storage_compat::EndpointServerPools;
-use super::storage_compat::TierConfigMgr;
-use super::storage_compat::metadata_sys::BucketMetadataSys;
+use super::super::storage_compat::EndpointServerPools;
+use super::super::storage_compat::TierConfigMgr;
+use super::super::storage_compat::metadata_sys::BucketMetadataSys;
 use crate::config::RustFSBufferConfig;
 use async_trait::async_trait;
 use rustfs_config::server_config::Config;
