+++ rustfs/src/storage/rpc/http_service.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::storage_compat::DEFAULT_READ_BUFFER_SIZE;
-use super::storage_compat::StorageDiskRpcExt as _;
-use super::storage_compat::WalkDirOptions;
-use super::storage_compat::find_local_disk_by_ref;
-use super::storage_compat::verify_rpc_signature;
+use super::super::storage_compat::DEFAULT_READ_BUFFER_SIZE;
+use super::super::storage_compat::StorageDiskRpcExt as _;
+use super::super::storage_compat::WalkDirOptions;
+use super::super::storage_compat::find_local_disk_by_ref;
+use super::super::storage_compat::verify_rpc_signature;
 use crate::server::RPC_PREFIX;
 use crate::storage::request_context::spawn_traced;
 use bytes::{Bytes, BytesMut};
